Move type_inference_visitor into python_ta package
