+++ python_ta/transforms/type_inference_visitor.py
