+++ setup.py
@@ -17,7 +17,7 @@ setup(
     author='David Liu',
     author_email='david@cs.toronto.edu',
     license='MIT',
-    packages=['python_ta', 'python_ta.reporters', 'python_ta.checkers', 'python_ta.docstring'],
+    packages=['python_ta', 'python_ta.reporters', 'python_ta.checkers', 'python_ta.docstring', 'python_ta.transforms'],
     install_requires=[
         'pycodestyle',
         'pylint',
